# Use memcpy() to dump fixnum/float