This patch uses standard C library to copy the string.
(Ref. https://github.com/ohler55/oj/pull/674)

−               | before | after  | result
--               | --     | --     | --
Oj.dump          | 1.046M | 1.102M | 1.054x

### Environment
- Zorin OS 16
- AMD Ryzen 7 5700G
- gcc version 11.1.0
- Ruby 3.1.0

### Before
```
Warming up --------------------------------------
             Oj.dump   106.035k i/100ms
Calculating -------------------------------------
             Oj.dump      1.046M (± 1.0%) i/s -     15.799M in  15.098842s
```

### After
```
Warming up --------------------------------------
             Oj.dump   112.786k i/100ms
Calculating -------------------------------------
             Oj.dump      1.102M (± 1.1%) i/s -     16.580M in  15.051956s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

data = {
  float: 3.141592653589793,
  fixnum: 2 ** 60
}

Benchmark.ips do |x|
  x.time = 15

  x.report('Oj.dump') { Oj.dump(data, mode: :compat) }
end
```